refactor(agent): factor out `handleBulkAgentOperation()`
