+++ .vscode/settings.json
@@ -20,5 +20,6 @@
   },
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": "never"
-  }
+  },
+  "vitest.disableWorkspaceWarning": true
 }
