+++ packages/server/src/fhir/operations/agentbulkstatus.ts
@@ -1,11 +1,7 @@
-import { allOk, badRequest, isOk, serverError } from '@medplum/core';
 import { FhirRequest, FhirResponse } from '@medplum/fhir-router';
-import { Agent, Bundle, BundleEntry, OperationDefinition, OperationOutcome, Parameters } from '@medplum/fhirtypes';
-import { getAuthenticatedContext } from '../../context';
+import { Agent, OperationDefinition } from '@medplum/fhirtypes';
 import { agentStatusHandler } from './agentstatus';
-import { getAgentsForRequest } from './agentutils';
-
-export const MAX_AGENTS_PER_PAGE = 100;
+import { handleBulkAgentOperation } from './agentutils';
 
 export const operation: OperationDefinition = {
   resourceType: 'OperationDefinition',
@@ -31,52 +27,7 @@ export const operation: OperationDefinition = {
  * @returns The FHIR response.
  */
 export async function agentBulkStatusHandler(req: FhirRequest): Promise<FhirResponse> {
-  const { repo } = getAuthenticatedContext();
-
-  if (req.query._count && Number.parseInt(req.query._count, 10) > MAX_AGENTS_PER_PAGE) {
-    return [badRequest(`'_count' of ${req.query._count} is greater than max of ${MAX_AGENTS_PER_PAGE}`)];
-  }
-
-  const agents = await getAgentsForRequest(req, repo);
-  if (!agents?.length) {
-    return [badRequest('No agent(s) for given query')];
-  }
-
-  const promises = agents.map((agent) => agentStatusHandler({ ...req, params: { id: agent.id as string } }));
-  const results = await Promise.allSettled(promises);
-  const entries = [] as BundleEntry<Parameters>[];
-  for (let i = 0; i < results.length; i++) {
-    const result = results[i];
-    if (result.status === 'rejected') {
-      entries.push(makeResultWrapperEntry(serverError(result.reason as Error), agents[i]));
-      continue;
-    }
-    const [outcome, params] = result.value;
-    if (!isOk(outcome)) {
-      entries.push(makeResultWrapperEntry(outcome, agents[i]));
-      continue;
-    }
-    entries.push(makeResultWrapperEntry(params as Parameters, agents[i]));
-  }
-
-  return [
-    allOk,
-    {
-      resourceType: 'Bundle',
-      type: 'collection',
-      entry: entries,
-    } satisfies Bundle<Parameters>,
-  ];
-}
-
-function makeResultWrapperEntry(result: Parameters | OperationOutcome, agent: Agent): BundleEntry<Parameters> {
-  return {
-    resource: {
-      resourceType: 'Parameters',
-      parameter: [
-        { name: 'agent', resource: agent },
-        { name: 'result', resource: result },
-      ],
-    },
-  };
+  return handleBulkAgentOperation(req, (agent: Agent) =>
+    agentStatusHandler({ ...req, params: { id: agent.id as string } })
+  );
 }
